// models/LicenseTemplateInfo.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * LicensesInfo contains information about a License
 */
export type LicenseTemplateInfo = {
    body?: string;
    implementation?: string;
    key?: string;
    name?: string;
    url?: string;
};